fix errors

+++ 16-binary_tree_is_perfect.c
@@ -41,7 +41,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
  */
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	int l, r;
+	size_t l, r;
 
 	if (!tree)
 		return (0);
